requirements.txt
django
djangorestframework
django-cors-headers
xarray
numpy
pandas
netcdf4
dask[complete]
plotly
